# Conflicts:
#	pom.xml
#	src/main/java/org/gmaoback/gmaomondi_backend/controllers/ArticleController.java
#	src/main/java/org/gmaoback/gmaomondi_backend/dao/entites/Article.java
#	src/main/java/org/gmaoback/gmaomondi_backend/dao/repositories/ArticleRepository.java

+++ src/main/java/org/gmaoback/gmaomondi_backend/dao/entites/Personne.java
@@ -20,7 +20,7 @@ public class Personne {
     private String nom;
     private String prenom;
     private String specialite;
-    private int tel;
+    private Long tel;
     private String email;
